+++ evals/report/pubsub/pubsub.go
@@ -26,6 +26,12 @@ var marshalOptions = protojson.MarshalOptions{
 	EmitUnpopulated: true,
 }
 
+// MarshalRunJSON serializes run using the Pub/Sub → BigQuery JSON contract.
+// Options must stay aligned with [Reporter.ReportRun] and bqschema assumptions.
+func MarshalRunJSON(run *evalspb.Run) ([]byte, error) {
+	return marshalOptions.Marshal(run)
+}
+
 // publisher is the write seam. A *pubsub.Publisher satisfies it via
 // realPublisher; tests substitute a fake.
 type publisher interface {
